refactor: 初步重构代码，剥离插件，开源许可证更改为MIT
