--- index.ts
@@ -1,7 +1,0 @@
-export { FloatingLive } from "./src";
-export * from "./src/types";
-export * from "./src/enum";
-
-export * from "./src/abstract/LiveRoom";
-export * from "./src/abstract/Reglist";
-export * from "./src/utils";
