--- plugin/chatPrint/index.ts
@@ -1,11 +1,0 @@
-import { FloatingLive } from "../..";
-import { chatPrint } from "./chatPrint";
-
-export = () => {
-  return {
-    name: "chatPrint",
-    register: (ctx: FloatingLive) => {
-      return new chatPrint(ctx);
-    },
-  };
-};
